added units of measurement. may be different between aentry and exit of goods. also added costs, which can be defined on acquisition of stock and derived on all of the different units of exit

+++ services/bime/src/main/java/bime/dto/StockAlertThresholdRequestDTO.java
@@ -3,6 +3,7 @@ package bime.dto;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
@@ -11,5 +12,5 @@ public class StockAlertThresholdRequestDTO {
     private UUID variantId;
     private UUID locationId;
     @Schema(description = "Alert fires the moment on-hand quantity at this location drops to or below this value", example = "10")
-    private int threshold;
+    private BigDecimal threshold;
 }
